requirements.txt
appdirs==1.4.3
crispy-forms-foundation==0.3.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==1.8.18
django-allauth==0.32.0
django-appconf==1.0.1
django-compressor==1.5
django-crispy-forms==1.6.1
django-eztables==0.3.2
django-generic-aggregation==0.3.2
django-geojson==2.10.0
django-google-analytics==1
django-leaflet==0.14.1
django-queryset-csv==1.0.0
django-tables2==1.5.0
django-timedeltafield==0.7.10
django.js==0.8.1
flickrapi==2.1.2
foundationform==0.2
future==0.16.0
html5lib==0.999
httplib2==0.9.2
oauth==1.0.1
oauth2==1.9.0.post1
oauthlib==2.0.2
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
PyPDF2==1.24
python-decouple==3.1
python-oauth2==1.0.0
python-openid==2.2.5
python-twitter==3.2.1
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
requests-toolbelt==0.6.0
six==1.10.0
tablib==0.10.0
unicodecsv==0.14.1
Unipath==1.1
